+++ scala-package/core/src/main/scala/org/apache/mxnet/NDArray.scala
@@ -562,30 +562,20 @@ object NDArray extends NDArrayBase {
  */
 class NDArray private[mxnet](private[mxnet] val handle: NDArrayHandle,
                              val writable: Boolean = true,
-                             addToCollector: Boolean = true)
-  extends WarnIfNotDisposed  // {
-     with NativeResource {
+                             addToCollector: Boolean = true) extends NativeResource {
   if (addToCollector) {
     NDArrayCollector.collect(this)
   }
 
-
   override def nativeAddress: CPtrAddress = handle
-
-  override def nativeDeAllocAddress: CPtrAddress => Int = _LIB.mxNDArrayFree
-
-  override val phantomRef: NativeResourceRef = super.register(this)
-
+  override def nativeDeAllocator: (CPtrAddress => Int) = _LIB.mxNDArrayFree
   override def bytesAllocated: Long = DType.numOfBytes(this.dtype) * this.shape.product
 
+  override val phantomRef: NativeResourceRef = super.register()
 
   // record arrays who construct this array instance
   // we use weak reference to prevent gc blocking
   private[mxnet] val dependencies = mutable.HashMap.empty[Long, WeakReference[NDArray]]
-
-//  @volatile private var disposed = false
-//  def isDisposed: Boolean = disposed
-  @volatile private var disposed = isDisposed
 
   def serialize(): Array[Byte] = {
     val buf = ArrayBuffer.empty[Byte]
@@ -598,15 +588,12 @@ class NDArray private[mxnet](private[mxnet] val handle: NDArrayHandle,
    * The NDArrays it depends on will NOT be disposed. <br />
    * The object shall never be used after it is disposed.
    */
-//  def dispose(): Unit = {
-//    print("dispose\n")
-//    if (!disposed) {
-//      print("disposing\n")
-//      _LIB.mxNDArrayFree(handle)
-//      dependencies.clear()
-//      disposed = true
-//    }
-//  }
+  override def dispose(): Unit = {
+    if (!disposed) {
+      super.dispose()
+      dependencies.clear()
+    }
+  }
 
   /**
    * Dispose all NDArrays who help to construct this array. <br />
