+++ scala-package/core/src/main/scala/org/apache/mxnet/Symbol.scala
@@ -29,33 +29,16 @@ import scala.collection.mutable.{ArrayBuffer, ListBuffer}
  * WARNING: it is your responsibility to clear this object through dispose().
  * </b>
  */
-class Symbol private(private[mxnet] val handle: SymbolHandle)
-  extends WarnIfNotDisposed with NativeResource {
+class Symbol private(private[mxnet] val handle: SymbolHandle) extends NativeResource {
+  private val logger: Logger = LoggerFactory.getLogger(classOf[Symbol])
+
+  // unable to get the byteAllocated for Symbol
+  override def bytesAllocated: Long = 0L
 
   override def nativeAddress: CPtrAddress = handle
+  override def nativeDeAllocator: (CPtrAddress => Int) = _LIB.mxSymbolFree
+  override val phantomRef: NativeResourceRef = super.register()
 
-  override def nativeDeAllocAddress: CPtrAddress => Int = _LIB.mxSymbolFree
-
-  override val phantomRef: NativeResourceRef = super.register(this)
-
-  override def bytesAllocated: Long = 0
-
-  private val logger: Logger = LoggerFactory.getLogger(classOf[Symbol])
-  private var disposed = false
-//  protected def isDisposed = disposed
-
-  /**
-   * Release the native memory.
-   * The object shall never be used after it is disposed.
-   */
-/*
-  def dispose(): Unit = {
-    if (!disposed) {
-      _LIB.mxSymbolFree(handle)
-      disposed = true
-    }
-  }
-*/
 
   def +(other: Symbol): Symbol = Symbol.createFromListedSymbols("_Plus")(Array(this, other))
   def +[@specialized(Int, Float, Double) V](other: V): Symbol = {
@@ -844,6 +827,7 @@ class Symbol private(private[mxnet] val handle: SymbolHandle)
     checkCall(_LIB.mxSymbolSaveToJSON(handle, jsonStr))
     jsonStr.value
   }
+
 }
 
 /**
